After restructuring code
