+++ .gitignore
@@ -1,5 +1,4 @@
 env
 .vscode
 /.project
-/.pydevproject
 twitter-secret-keys.yaml
